Updated JWT and protected endpoint, basic test compliant, need to be tested further

+++ part3/app/api/v1/reviews.py
@@ -1,5 +1,6 @@
 from flask_restx import Namespace, Resource, fields
 from app.services import facade
+from flask_jwt_extended import jwt_required, get_jwt_identity
 
 api = Namespace('reviews', description='Review operations')
 
@@ -16,28 +17,26 @@ class ReviewList(Resource):
     @api.expect(review_model)
     @api.response(201, 'Review successfully created')
     @api.response(400, 'Invalid input data')
+    @jwt_required()
     def post(self):
         """Register a new review"""
         review_data = api.payload
-        user = facade.get_user(review_data['user_id'])
         place = facade.get_place(review_data['place_id'])
-
-        if not user or not place:
-            return {'error': 'Invalid User or Place ID'}, 400
-
-        if place.owner_id == user.id:
-            return {'error': 'User cannot review their own place'}, 400
-
+        if not place:
+            return {'error': 'Place not found'}, 400
+        user = facade.get_user(review_data['user_id'])
+        if not user:
+            return {'error': 'User not found'}, 400
+        if place.owner.id == user.id:
+            return {'error': 'You cannot review your own place'}, 400
+        existing_review = facade.get_review_by_place_and_user(review_data['place_id'], review_data['user_id'])
+        if existing_review:
+            return {'error': 'You have already reviewed this place'}, 400
         try:
             new_review = facade.create_review(review_data)
             return new_review.to_dict(), 201
         except Exception as e:
             return {'error': str(e)}, 400
-
-    @api.response(200, 'List of reviews retrieved successfully')
-    def get(self):
-        """Retrieve a list of all reviews"""
-        return [review.to_dict() for review in facade.get_all_reviews()], 200
 
 
     @api.response(200, 'List of reviews retrieved successfully')
@@ -60,13 +59,16 @@ class ReviewResource(Resource):
     @api.response(200, 'Review updated successfully')
     @api.response(404, 'Review not found')
     @api.response(400, 'Invalid input data')
+    @jwt_required()
     def put(self, review_id):
         """Update a review's information"""
         review_data = api.payload
-        review = facade.get_review(review_id)
+        review = facade.get_review_by_id(review_id)
         if not review:
             return {'error': 'Review not found'}, 404
-        
+        current_user_id = get_jwt_identity()
+        if review.user_id != current_user_id:
+            return {'error': 'Unauthorized action.'}, 403
         try:
             facade.update_review(review_id, review_data)
             return {'message': 'Review updated successfully'}, 200
@@ -75,12 +77,15 @@ class ReviewResource(Resource):
 
     @api.response(200, 'Review deleted successfully')
     @api.response(404, 'Review not found')
+    @jwt_required
     def delete(self, review_id):
         """Delete a review"""
         review = facade.get_review(review_id)
         if not review:
             return {'error': 'Review not found'}, 404
-        
+        current_user_id = get_jwt_identity()
+        if review.user_id != current_user_id:
+            return {'error': 'Unauthorized action.'}, 403
         try:
             facade.delete_review(review_id)
             return {'message': 'Review deleted successfully'}, 200
